write cambiado y ordenadas las funciones y librerias de Kernel

+++ RowDaBoat-x64barebones-d4e1c147f975/Kernel/syscalls/cpuinfo.c
@@ -10,5 +10,5 @@ uint64_t sys_cpuinfo(char buffer[70])
 
 	cpuBrand(&(buffer[16]));
 
-	return buffer;
+	return (uint64_t) buffer;
 }
